+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `unique-by` for more control over how uniqueness is determined.
 - `for` as a better pattern for doing something effectful over the stream.
+- `any?`, `all?`, and `reduced?` as modern aliases.
 
 #### Deprecated