+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Transducers
 
-### Unreleased
+### 1.4.0 (2025-02-15)
 
 #### Added